surrealdb_types::duration

Struct Duration

pub struct Duration(/* private fields */);

Represents a duration value in SurrealDB

A duration represents a span of time, typically used for time-based calculations and comparisons. This type wraps the standard std::time::Duration type.

Implementations

impl Duration

pub const MAX: Duration

The maximum duration

pub const ZERO: Duration

The zero duration

pub fn new(secs: u64, nanos: u32) -> Duration

Create a duration from both seconds and nanoseconds components

pub fn from_std(d: Duration) -> Self

Create a duration from std::time::Duration

pub fn into_inner(self) -> Duration

Convert into the inner std::time::Duration

pub fn nanos(&self) -> u128

Get the total number of nanoseconds

pub fn micros(&self) -> u128

Get the total number of microseconds

pub fn millis(&self) -> u128

Get the total number of milliseconds

pub fn secs(&self) -> u64

Get the total number of seconds

pub fn mins(&self) -> u64

Get the total number of minutes

pub fn hours(&self) -> u64

Get the total number of hours

pub fn days(&self) -> u64

Get the total number of dats

pub fn weeks(&self) -> u64

Get the total number of months

pub fn years(&self) -> u64

Get the total number of years

pub fn from_nanos(nanos: u64) -> Duration

Create a duration from nanoseconds

pub fn from_micros(micros: u64) -> Duration

Create a duration from microseconds

pub fn from_millis(millis: u64) -> Duration

Create a duration from milliseconds

pub fn from_secs(secs: u64) -> Duration

Create a duration from seconds

pub fn from_mins(mins: u64) -> Option<Duration>

Create a duration from minutes

pub fn from_hours(hours: u64) -> Option<Duration>

Create a duration from hours

pub fn from_days(days: u64) -> Option<Duration>

Create a duration from days

pub fn from_weeks(weeks: u64) -> Option<Duration>

Create a duration from weeks