surrealdb_types

Enum Geometry

pub enum Geometry {
    Point(Point<f64>),
    Line(LineString<f64>),
    Polygon(Polygon<f64>),
    MultiPoint(MultiPoint<f64>),
    MultiLine(MultiLineString<f64>),
    MultiPolygon(MultiPolygon<f64>),
    Collection(Vec<Geometry>),
}

Represents geometric shapes in SurrealDB

Geometry types support various geometric shapes including points, lines, polygons, and their multi-variants. This is useful for spatial data and geographic applications.

The types used internally originate from the geo crate.

Variants

Point(Point<f64>)

A single point in 2D space

Line(LineString<f64>)

A line consisting of multiple connected points

Polygon(Polygon<f64>)

A polygon with an exterior boundary and optional interior holes

MultiPoint(MultiPoint<f64>)

Multiple points

MultiLine(MultiLineString<f64>)

Multiple lines

MultiPolygon(MultiPolygon<f64>)

Multiple polygons

Collection(Vec<Geometry>)

A collection of different geometry types

Implementations

impl Geometry

pub fn kind(&self) -> GeometryKind

Get the kind of geometry

pub fn is_point(&self) -> bool

Check if this is a of the given type

pub fn into_point(self) -> Result<Point<f64>>

Convert this geometry into the given type

pub fn from_point(v: Point<f64>) -> Self

Create a new geometry from the given type

pub fn is_line(&self) -> bool

Check if this is a of the given type

pub fn into_line(self) -> Result<LineString<f64>>

Convert this geometry into the given type

pub fn from_line(v: LineString<f64>) -> Self

Create a new geometry from the given type

pub fn is_polygon(&self) -> bool

Check if this is a of the given type

pub fn into_polygon(self) -> Result<Polygon<f64>>

Convert this geometry into the given type

pub fn from_polygon(v: Polygon<f64>) -> Self

Create a new geometry from the given type

pub fn is_multipoint(&self) -> bool

Check if this is a of the given type

pub fn into_multipoint(self) -> Result<MultiPoint<f64>>

Convert this geometry into the given type

pub fn from_multipoint(v: MultiPoint<f64>) -> Self

Create a new geometry from the given type

pub fn is_multiline(&self) -> bool

Check if this is a of the given type

pub fn into_multiline(self) -> Result<MultiLineString<f64>>

Convert this geometry into the given type

pub fn from_multiline(v: MultiLineString<f64>) -> Self

Create a new geometry from the given type

pub fn is_multipolygon(&self) -> bool

Check if this is a of the given type

pub fn into_multipolygon(self) -> Result<MultiPolygon<f64>>

Convert this geometry into the given type

pub fn from_multipolygon(v: MultiPolygon<f64>) -> Self

Create a new geometry from the given type

pub fn is_collection(&self) -> bool

Check if this is a of the given type

pub fn into_collection(self) -> Result<Vec<Geometry>>

Convert this geometry into the given type

pub fn from_collection(v: Vec<Geometry>) -> Self

Create a new geometry from the given type

impl Geometry

pub fn is_valid(&self) -> bool

Check if this has valid latitude and longitude points:

-90 <= lat <= 90

-180 <= lng <= 180

pub fn as_type(&self) -> &'static str

Get the type of this Geometry as text

pub fn as_coordinates(&self) -> Value

Get the raw coordinates of this Geometry as an Array

pub fn as_object(&self) -> Object

Convert this geometry into an object

pub fn try_from_object(object: &Object) -> Option<Geometry>

Try to convert an object into a geometry