surrealdb_types

Enum Value

pub enum Value {

    None,
    Null,
    Bool(bool),
    Number(Number),
    String(String),
    Bytes(Bytes),
    Duration(Duration),
    Datetime(Datetime),
    Uuid(Uuid),
    Geometry(Geometry),
    Table(Table),
    RecordId(RecordId),
    File(File),
    Range(Box<Range>),
    Regex(Regex),
    Array(Array),
    Object(Object),
    Set(Set),

}

Represents a value in SurrealDB

This enum contains all possible value types that can be stored in SurrealDB. Each variant corresponds to a different data type supported by the database.

Variants

None

Represents the absence of a value

Null

Represents a null value

Bool(bool)

A boolean value (true or false)

Number(Number)

A numeric value (integer, float, or decimal)

String(String)

A string value

Bytes(Bytes)

Binary data

Duration(Duration)

A duration value representing a time span

Datetime(Datetime)

A datetime value representing a point in time

Uuid(Uuid)

A UUID value

Geometry(Geometry)

A geometric value (point, line, polygon, etc.)

Table(Table)

A table value

RecordId(RecordId)

A record identifier

File(File)

A file reference

Range(Box<Range>)

A range of values

Regex(Regex)

A regular expression

Array(Array)

An array of values

Object(Object)

An object containing key-value pairs

Set(Set)

A set of values

Implementations

impl Value

pub fn is_none(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_null(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_bool(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_bool(value: bool) -> Value

Converts the given value into a Value

impl Value

pub fn into_bool(self) -> Result<bool, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_bool(&self) -> Option<&bool>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_true(&self) -> bool

Checks if this value is specifically true

Returns true if the value is Value::Bool(true), false otherwise.

pub fn is_false(&self) -> bool

Checks if this value is specifically false

Returns true if the value is Value::Bool(false), false otherwise.

impl Value

pub fn is_number(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_number(value: Number) -> Value

Converts the given value into a Value

impl Value

pub fn into_number(self) -> Result<Number, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_number(&self) -> Option<&Number>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_number_and(&self, callback: impl FnOnce(&Number) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_int(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_i64(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_int(value: i64) -> Value

Converts the given value into a Value

impl Value

pub fn from_i64(value: i64) -> Value

Converts the given value into a Value

impl Value

pub fn into_int(self) -> Result<i64, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn into_i64(self) -> Result<i64, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_int(&self) -> Option<&i64>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

pub fn as_i64(&self) -> Option<&i64>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_int_and(&self, callback: impl FnOnce(&i64) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

pub fn is_i64_and(&self, callback: impl FnOnce(&i64) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_float(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_f64(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_float(value: f64) -> Value

Converts the given value into a Value

impl Value

pub fn from_f64(value: f64) -> Value

Converts the given value into a Value

impl Value

pub fn into_float(self) -> Result<f64, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn into_f64(self) -> Result<f64, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_float(&self) -> Option<&f64>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

pub fn as_f64(&self) -> Option<&f64>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_float_and(&self, callback: impl FnOnce(&f64) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

pub fn is_f64_and(&self, callback: impl FnOnce(&f64) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_decimal(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_decimal(value: Decimal) -> Value

Converts the given value into a Value

impl Value

pub fn into_decimal(self) -> Result<Decimal, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_decimal(&self) -> Option<&Decimal>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_decimal_and(&self, callback: impl FnOnce(&Decimal) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_string(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_string(value: String) -> Value

Converts the given value into a Value

impl Value

pub fn into_string(self) -> Result<String, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_string(&self) -> Option<&String>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_string_and(&self, callback: impl FnOnce(&String) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_duration(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_duration(value: Duration) -> Value

Converts the given value into a Value

impl Value

pub fn into_duration(self) -> Result<Duration, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_duration(&self) -> Option<&Duration>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_duration_and(&self, callback: impl FnOnce(&Duration) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_datetime(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_datetime(value: Datetime) -> Value

Converts the given value into a Value

impl Value

pub fn into_datetime(self) -> Result<Datetime, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_datetime(&self) -> Option<&Datetime>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_datetime_and(&self, callback: impl FnOnce(&Datetime) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_uuid(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_uuid(value: Uuid) -> Value

Converts the given value into a Value

impl Value

pub fn into_uuid(self) -> Result<Uuid, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_uuid(&self) -> Option<&Uuid>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_uuid_and(&self, callback: impl FnOnce(&Uuid) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_array(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_array(value: Array) -> Value

Converts the given value into a Value

impl Value

pub fn into_array(self) -> Result<Array, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_array(&self) -> Option<&Array>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_array_and(&self, callback: impl FnOnce(&Array) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_set(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_set(value: Set) -> Value

Converts the given value into a Value

impl Value

pub fn into_set(self) -> Result<Set, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_set(&self) -> Option<&Set>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_set_and(&self, callback: impl FnOnce(&Set) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_object(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_object(value: Object) -> Value

Converts the given value into a Value

impl Value

pub fn into_object(self) -> Result<Object, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_object(&self) -> Option<&Object>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_object_and(&self, callback: impl FnOnce(&Object) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_geometry(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_geometry(value: Geometry) -> Value

Converts the given value into a Value

impl Value

pub fn into_geometry(self) -> Result<Geometry, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_geometry(&self) -> Option<&Geometry>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_geometry_and(&self, callback: impl FnOnce(&Geometry) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_bytes(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_bytes(value: Bytes) -> Value

Converts the given value into a Value

impl Value

pub fn into_bytes(self) -> Result<Bytes, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_bytes(&self) -> Option<&Bytes>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_bytes_and(&self, callback: impl FnOnce(&Bytes) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_table(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_table(value: Table) -> Value

Converts the given value into a Value

impl Value

pub fn into_table(self) -> Result<Table, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_table(&self) -> Option<&Table>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_table_and(&self, callback: impl FnOnce(&Table) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_record(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_record(value: RecordId) -> Value

Converts the given value into a Value

impl Value

pub fn into_record(self) -> Result<RecordId, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_record(&self) -> Option<&RecordId>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_record_and(&self, callback: impl FnOnce(&RecordId) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_file(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_file(value: File) -> Value

Converts the given value into a Value

impl Value

pub fn into_file(self) -> Result<File, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_file(&self) -> Option<&File>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_file_and(&self, callback: impl FnOnce(&File) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_range(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_range(value: Range) -> Value

Converts the given value into a Value

impl Value

pub fn into_range(self) -> Result<Range, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_range(&self) -> Option<&Range>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_range_and(&self, callback: impl FnOnce(&Range) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_vec<T: SurrealValue>(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_vec<T: SurrealValue>(value: Vec<T>) -> Value

Converts the given value into a Value

impl Value

pub fn into_vec<T: SurrealValue>(self) -> Result<Vec<T>, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn is_option<T: SurrealValue>(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_option<T: SurrealValue>(value: Option<T>) -> Value

Converts the given value into a Value

impl Value

pub fn into_option<T: SurrealValue>(self) -> Result<Option<T>, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn from_btreemap<V: SurrealValue>(value: BTreeMap<String, V>) -> Value

Converts the given value into a Value

impl Value

pub fn into_btreemap<V: SurrealValue>( self, ) -> Result<BTreeMap<String, V>, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn from_hashmap<V: SurrealValue>(value: HashMap<String, V>) -> Value

Converts the given value into a Value

impl Value

pub fn into_hashmap<V: SurrealValue>(self) -> Result<HashMap<String, V>, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn is_point(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_point(value: Point) -> Value

Converts the given value into a Value

impl Value

pub fn into_point(self) -> Result<Point, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_point(&self) -> Option<&Point>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_point_and(&self, callback: impl FnOnce(&Point) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_line(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_line(value: LineString) -> Value

Converts the given value into a Value

impl Value

pub fn into_line(self) -> Result<LineString, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_line(&self) -> Option<&LineString>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_line_and(&self, callback: impl FnOnce(&LineString) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_polygon(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_polygon(value: Polygon) -> Value

Converts the given value into a Value

impl Value

pub fn into_polygon(self) -> Result<Polygon, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_polygon(&self) -> Option<&Polygon>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_polygon_and(&self, callback: impl FnOnce(&Polygon) -> bool) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_multipoint(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_multipoint(value: MultiPoint) -> Value

Converts the given value into a Value

impl Value

pub fn into_multipoint(self) -> Result<MultiPoint, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_multipoint(&self) -> Option<&MultiPoint>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_multipoint_and( &self, callback: impl FnOnce(&MultiPoint) -> bool, ) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_multiline(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_multiline(value: MultiLineString) -> Value

Converts the given value into a Value

impl Value

pub fn into_multiline(self) -> Result<MultiLineString, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_multiline(&self) -> Option<&MultiLineString>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_multiline_and( &self, callback: impl FnOnce(&MultiLineString) -> bool, ) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_multipolygon(&self) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn from_multipolygon(value: MultiPolygon) -> Value

Converts the given value into a Value

impl Value

pub fn into_multipolygon(self) -> Result<MultiPolygon, Error>

Attempts to convert a SurrealDB value into the given type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

impl Value

pub fn as_multipolygon(&self) -> Option<&MultiPolygon>

Returns a reference to the inner value if it matches the expected type.

Returns Some(&T) if the value is of the expected type, None otherwise.

impl Value

pub fn is_multipolygon_and( &self, callback: impl FnOnce(&MultiPolygon) -> bool, ) -> bool

Checks if this value can be converted to the given type

Returns true if the value can be converted to the type, false otherwise.

impl Value

pub fn is_i8(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_i8(value: i8) -> Value

Converts the given value into a Value

pub fn into_i8(self) -> Result<i8, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_i8_and(&self, callback: impl FnOnce(&i8) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_i8(&self) -> Option<i8>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_i16(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_i16(value: i16) -> Value

Converts the given value into a Value

pub fn into_i16(self) -> Result<i16, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_i16_and(&self, callback: impl FnOnce(&i16) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_i16(&self) -> Option<i16>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_i32(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_i32(value: i32) -> Value

Converts the given value into a Value

pub fn into_i32(self) -> Result<i32, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_i32_and(&self, callback: impl FnOnce(&i32) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_i32(&self) -> Option<i32>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_isize(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_isize(value: isize) -> Value

Converts the given value into a Value

pub fn into_isize(self) -> Result<isize, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_isize_and(&self, callback: impl FnOnce(&isize) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_isize(&self) -> Option<isize>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_u8(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_u8(value: u8) -> Value

Converts the given value into a Value

pub fn into_u8(self) -> Result<u8, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_u8_and(&self, callback: impl FnOnce(&u8) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_u8(&self) -> Option<u8>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_u16(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_u16(value: u16) -> Value

Converts the given value into a Value

pub fn into_u16(self) -> Result<u16, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_u16_and(&self, callback: impl FnOnce(&u16) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_u16(&self) -> Option<u16>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_u32(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_u32(value: u32) -> Value

Converts the given value into a Value

pub fn into_u32(self) -> Result<u32, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_u32_and(&self, callback: impl FnOnce(&u32) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_u32(&self) -> Option<u32>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_u64(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_u64(value: u64) -> Value

Converts the given value into a Value

pub fn into_u64(self) -> Result<u64, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_u64_and(&self, callback: impl FnOnce(&u64) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_u64(&self) -> Option<u64>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_usize(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_usize(value: usize) -> Value

Converts the given value into a Value

pub fn into_usize(self) -> Result<usize, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_usize_and(&self, callback: impl FnOnce(&usize) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_usize(&self) -> Option<usize>

Returns a reference-like value if it matches the expected type Note: For integer types, this returns None since the value needs conversion

impl Value

pub fn is_f32(&self) -> bool

Checks if this value can be converted to the given type

pub fn from_f32(value: f32) -> Value

Converts the given value into a Value

pub fn into_f32(self) -> Result<f32, Error>

Attempts to convert a SurrealDB value into the given type

pub fn is_f32_and(&self, callback: impl FnOnce(&f32) -> bool) -> bool

Checks if this value matches the type and passes the callback check

pub fn as_f32(&self) -> Option<f32>

Returns the converted value if it matches the expected type Note: For float types, this returns a converted value, not a reference

impl Value

pub fn into_json_value(self) -> JsonValue

Converts the value into a json representation of the value. Returns None if there are non serializable values present in the value.

impl Value

pub fn kind(&self) -> Kind

Returns the kind of this value

This method maps each value variant to its corresponding Kind.

pub fn first(&self) -> Option<Value>

Returns the first value in the array.

Returns None if the value is not an array or the array is empty.

pub fn is_nullish(&self) -> bool

Check if this Value is NONE or NULL

pub fn is_empty(&self) -> bool

Check if this Value is empty.

pub fn get<Idx>(&self, index: Idx) -> &Value
where Value: Indexable<Idx>,

Accesses the value found at a certain field if an object, and a certain index if an array. Will not err if no value is found at this point, instead returning a Value::None. If an Option<&Value> is desired, the .into_option() method can be used to perform the conversion.

pub fn remove<Idx>(&mut self, index: Idx) -> Value
where Value: Indexable<Idx>,

Removes the value at the given index.

Returns Some(Value) if the value was removed, None if the value was not found.

pub fn is<T: SurrealValue>(&self) -> bool

Checks if this value is of the specified type

Returns true if the value can be converted to the given type T.

pub fn into_t<T: SurrealValue>(self) -> Result<T, Error>

Converts this value to the specified type

Returns Ok(T) if the conversion is successful, Err(Error) otherwise.

pub fn from_t<T: SurrealValue>(value: T) -> Value

Creates a value from the specified type

Converts the given value of type T into a Value.

pub fn is_kind(&self, kind: &Kind) -> bool

Check if this value matches the specified kind

Returns true if the value conforms to the given kind specification. This includes checking nested types for arrays, objects, records, etc.