surrealdb_types

Module record_id

Record identifier value types for SurrealDB

Modules

key

Record id key types

range

Record id range types

Structs

RecordId

Represents a record identifier in SurrealDB

RecordIdKeyRange

Represents a range of record identifier keys in SurrealDB

Enums

RecordIdKey

Represents a key component of a record identifier in SurrealDB

Constants

ID_CHARS

The characters which are supported in server record IDs